chore(ruff): bump 0.15.17 -> 0.16.5 and take the mechanical fixes (#1157)
Bumps ruff in .pre-commit-config.yaml and in the `make lint-tools` pin, so the
local gate and the CI hooks stay in lockstep.

611 findings across 158 files are fixed here. 509 had safe fixes. The other 102
came from --unsafe-fixes, applied ONLY for rules where the transformation is
mechanical and provably behaviour-preserving:

  RUF046  redundant int() around round() / math.ceil(), both of which already
          return int on Python 3
  RUF059  unused unpacked bindings
  PIE810, C408, C409, FLY002, ISC004, SIM*, PLC0206, DTZ011, RUF007, RUF022

black and isort were re-run afterwards so the three formatters agree;
`pre-commit run --all-files` reaches a fixed point (verified: two consecutive
runs leave an identical tree).

34 findings are DEFERRED, not fixed, each with a rationale in the
[tool.ruff.lint] ignore block of the root, ai/ and tools/vmaf-tune/ configs --
ruff resolves config per-file from the nearest pyproject.toml, so all three
needed the entry. The intent is to keep this gate exactly as strict as it was
BEFORE the bump: a version upgrade should not smuggle in a refactor. None of
the 34 flags a defect we know of. Two worth calling out:

  PLR0917  too-many-positional-arguments -- the positional-only variant of
           PLR0913, which this config already ignores with the note that "VMAF
           feature kernels naturally take many". Same fact, same treatment.
  BLE001   blind-except -- the batch drivers under ai/scripts/ catch Exception
           per item deliberately, so one bad row does not kill a multi-hour
           run, and predictor.py does the same to fall back to the analytical
           curve when onnxruntime is absent, mirroring the Python original.

One dead end worth recording: mid-way through, `pre-commit run --all-files`
appeared not to converge, with isort and ruff seemingly rewriting each other. I
removed "I" from ruff's select on the theory that two tools were fighting over
import order. That was wrong -- isort reports "Fixing <file>" while writing
byte-identical content, and pre-commit flags the mtime. git saw no change at
all. The select entry is restored.

+++ tools/vmaf-tune/tests/test_workdir_enospc.py
@@ -33,7 +33,7 @@ import pytest
 _HERE = Path(__file__).resolve().parent
 sys.path.insert(0, str(_HERE.parent / "src"))
 
-from vmaftune.bisect import (  # noqa: E402
+from vmaftune.bisect import (
     _check_disk_space,
     _estimate_yuv_bytes,
     _workdir_parent,
@@ -138,7 +138,7 @@ class TestCheckDiskSpace:
 
         est = 100 * 1024**3
         # required = ceil(est * 1.1) — exactly what _check_disk_space computes
-        required = int(math.ceil(est * 1.1))
+        required = math.ceil(est * 1.1)
         with unittest.mock.patch(
             "vmaftune.bisect.shutil.disk_usage",
             return_value=_DiskUsage(total=required, used=0, free=required),
